// web/src/app/types/types.ts
export interface Episode {
    _id: string;
    air_date: Date;
    duration: number;
    eid: number | string;
    full_titles: string[];
    title: string;
    anime_id: string;
    viewed: string[];
    view: boolean;
    iframe: string;
}
// ...
export interface IAnime {
    _id: string;
    aid: number;
    bookmarks: string[];
    country?: string;
    episode: Episode[];
    episodes: number;
    title: string;
    full_title: string;
    image: string;
    title_en: string;
    title_jp?: string;
    quality: number;
    rating: number;
    release_date: Date;
    subscribe: boolean;
    subscribers?: string[];
    time?: number;
    url: string;
    description: string[];
    tags: string[]
}
// ...
export interface TabComponent{
    component: JSX.Element;
}

/** Tab Style
 * @property {string} label - `Tab Name`
 * @property {JSX.Element} icon - `Tab Render Icon`
 */
export interface TabProps{
    label: string;
    icon: JSX.Element;
}